backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1

# Database
psycopg[binary]==3.2.1
pgvector==0.3.2


# LangChain stack
langchain==0.3.5
langchain-community==0.3.3
langchain-openai==0.2.3
langchain-text-splitters==0.3.0

# Utils
requests==2.32.3

# PDF Processing
PyPDF2==3.0.1
python-multipart==0.0.9
tiktoken==0.9.0
